refactor(fe): integrate grade page into assignment page
