+++ apps/frontend/app/admin/course/[courseId]/assignment/[assignmentId]/(overall)/@tabs/page.tsx
@@ -1,20 +1,33 @@
+'use client'
+
 import { FetchErrorFallback } from '@/components/FetchErrorFallback'
+import { KatexContent } from '@/components/KatexContent'
+import { GET_ASSIGNMENT } from '@/graphql/assignment/queries'
+import { useQuery } from '@apollo/client'
 import { ErrorBoundary } from '@suspensive/react'
 import { Suspense } from 'react'
-import {
-  ParticipantTable,
-  ParticipantTableFallback
-} from './_components/ParticipantTable'
+import { ParticipantTableFallback } from './_components/ParticipantTable'
 
-export default function Submission({
-  params
-}: {
-  params: { assignmentId: string }
-}) {
+interface InformationProps {
+  params: { courseId: string; assignmentId: string }
+}
+
+export default function Information({ params }: InformationProps) {
+  const assignmentData = useQuery(GET_ASSIGNMENT, {
+    variables: {
+      groupId: Number(params.courseId),
+      assignmentId: Number(params.assignmentId)
+    }
+  }).data?.getAssignment
   return (
     <ErrorBoundary fallback={FetchErrorFallback}>
       <Suspense fallback={<ParticipantTableFallback />}>
-        <ParticipantTable assignmentId={Number(params.assignmentId)} />
+        <div>
+          <KatexContent
+            content={assignmentData?.description}
+            classname="mb-4 w-full max-w-full border-y-gray-300"
+          />
+        </div>
       </Suspense>
     </ErrorBoundary>
   )
