+++ apps/frontend/app/admin/course/[courseId]/assignment/[assignmentId]/_components/AssignmentOverallTabs.tsx
@@ -2,41 +2,26 @@
 
 import {
   GET_ASSIGNMENT_SCORE_SUMMARIES,
-  GET_ASSIGNMENT_SUBMISSION_SUMMARIES_OF_USER,
-  GET_ASSIGNMENTS
+  GET_ASSIGNMENT_SUBMISSION_SUMMARIES_OF_USER
 } from '@/graphql/assignment/queries'
-import { GET_ASSIGNMENT_PROBLEMS } from '@/graphql/problem/queries'
 import { cn } from '@/libs/utils'
-import excelIcon from '@/public/icons/excel.svg'
 import { useQuery } from '@apollo/client'
-import type { Route } from 'next'
-import Image from 'next/image'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
-import { CSVLink } from 'react-csv'
 
-interface ScoreSummary {
-  username: string
-  realName: string
-  studentId: string
-  userAssignmentScore: number
-  assignmentPerfectScore: number
-  submittedProblemCount: number
-  totalProblemCount: number
-  problemScores: { problemId: number; score: number; maxScore: number }[]
-  major: string
-}
 interface SubmissionSummary {
   problemId: number
+}
+
+interface AssignmentOverallTabsProps {
+  groupId: number
+  assignmentId: number
 }
 
 export function AssignmentOverallTabs({
   groupId,
   assignmentId
-}: {
-  groupId: number
-  assignmentId: number
-}) {
+}: AssignmentOverallTabsProps) {
   const pathname = usePathname()
 
   const { data: userData } = useQuery<{
@@ -47,99 +32,11 @@ export function AssignmentOverallTabs({
 
   const userId = userData?.getUserIdByAssignment?.userId
 
-  const { data: scoreData } = useQuery<{
-    getAssignmentScoreSummaries: ScoreSummary[]
-  }>(GET_ASSIGNMENT_SCORE_SUMMARIES, {
-    variables: { groupId, assignmentId, take: 300 }
-  })
-
   useQuery<{
     getAssignmentSubmissionSummaryByUserId: { submissions: SubmissionSummary[] }
   }>(GET_ASSIGNMENT_SUBMISSION_SUMMARIES_OF_USER, {
     variables: { groupId, assignmentId, userId, take: 300 }
   })
-
-  const { data: assignmentData } = useQuery(GET_ASSIGNMENTS, {
-    variables: { groupId, take: 100 }
-  })
-
-  const { data: problemData } = useQuery(GET_ASSIGNMENT_PROBLEMS, {
-    variables: { groupId, assignmentId }
-  })
-
-  const formatScore = (score: number): string => {
-    const fixedScore = Math.floor(score * 1000) / 1000
-    return fixedScore.toString()
-  }
-  const assignmentTitle = assignmentData?.getAssignments.find(
-    (assignment) => assignment.id === assignmentId.toString()
-  )?.title
-
-  const fileName = assignmentTitle
-    ? `${assignmentTitle.replace(/\s+/g, '_')}.csv`
-    : `course-${groupId}/assignment-${assignmentId}-participants.csv`
-
-  const problemList =
-    problemData?.getAssignmentProblems
-      .slice()
-      .sort((a, b) => a.order - b.order)
-      .map((problem) => ({
-        problemId: problem.problemId,
-        maxScore: problem.score,
-        title: problem.problem.title,
-        order: problem.order
-      })) || []
-
-  const problemHeaders = problemList.map((problem, index) => {
-    const problemLabel = String.fromCharCode(65 + index)
-    return {
-      label: `${problemLabel}(배점 ${problem.maxScore})`,
-      key: `problems[${index}].maxScore`
-    }
-  })
-
-  const headers = [
-    { label: '학번', key: 'studentId' },
-
-    { label: '전공', key: 'major' },
-    { label: '이름', key: 'realName' },
-    { label: '아이디', key: 'username' },
-    {
-      label: `제출 문제 수(총 ${scoreData?.getAssignmentScoreSummaries[0]?.totalProblemCount || 0})`,
-      key: 'problemRatio'
-    },
-    {
-      label: `총 획득 점수(만점 ${scoreData?.getAssignmentScoreSummaries[0]?.assignmentPerfectScore || 0})`,
-      key: 'score'
-    },
-
-    ...problemHeaders
-  ]
-
-  const csvData =
-    scoreData?.getAssignmentScoreSummaries.map((user) => {
-      const userProblemScores = problemList.map((problem) => {
-        const scoreData = user.problemScores.find(
-          (ps) => ps.problemId === problem.problemId
-        )
-
-        return {
-          maxScore: scoreData ? formatScore(scoreData.score) : '-'
-        }
-      })
-
-      return {
-        studentId: user.studentId,
-        major: user.major,
-        realName: user.realName,
-        username: user.username,
-        problemRatio: user.submittedProblemCount
-          ? `${user.submittedProblemCount}`
-          : '-',
-        score: formatScore(user.userAssignmentScore),
-        problems: userProblemScores
-      }
-    }) || []
 
   const isCurrentTab = (tab: string) => {
     if (tab === '') {
@@ -151,44 +48,53 @@ export function AssignmentOverallTabs({
   }
 
   return (
-    <div className="flex items-center justify-between">
-      <div className="flex w-max gap-1 rounded-lg bg-slate-200 p-1">
-        <Link
-          href={`/admin/course/${groupId}/assignment/${assignmentId}` as Route}
-          className={cn(
-            'rounded-md px-3 py-1.5 text-lg font-semibold',
-            isCurrentTab('') && 'text-primary bg-white font-bold'
-          )}
-        >
-          Participant
-        </Link>
-        <Link
-          href={
-            `/admin/course/${groupId}/assignment/${assignmentId}/submission` as Route
-          }
-          className={cn(
-            'rounded-md px-3 py-1.5 text-lg font-semibold',
-            isCurrentTab('submission') && 'text-primary bg-white font-bold'
-          )}
-        >
-          All Submission
-        </Link>
-      </div>
-      <CSVLink
-        data={csvData}
-        headers={headers}
-        filename={fileName}
-        className="flex items-center gap-2 rounded-lg bg-blue-400 px-3 py-1.5 text-lg font-semibold text-white transition-opacity hover:opacity-85"
+    <div className="flex h-[48px] w-full rounded-full border border-solid border-[#80808040] bg-white text-lg font-normal text-[#737373]">
+      <Link
+        href={`/admin/course/${groupId}/assignment/${assignmentId}` as const}
+        className={cn(
+          'flex h-full w-1/4 items-center justify-center rounded-full',
+          isCurrentTab('') &&
+            'text-primary border-primary border-2 border-solid bg-white'
+        )}
       >
-        Export
-        <Image
-          src={excelIcon}
-          alt="Excel Icon"
-          width={20}
-          height={20}
-          className="ml-1"
-        />
-      </CSVLink>
+        INFORMATION
+      </Link>
+      <Link
+        href={
+          `/admin/course/${groupId}/assignment/${assignmentId}/submission` as const
+        }
+        className={cn(
+          'flex h-full w-1/4 items-center justify-center rounded-full',
+          isCurrentTab('submission') &&
+            'text-primary border-primary border-2 border-solid bg-white'
+        )}
+      >
+        SUBMISSION
+      </Link>
+      <Link
+        href={
+          `/admin/course/${groupId}/assignment/${assignmentId}/assessment` as const
+        }
+        className={cn(
+          'flex h-full w-1/4 items-center justify-center rounded-full',
+          isCurrentTab('assessment') &&
+            'text-primary border-primary border-2 border-solid bg-white'
+        )}
+      >
+        ASSESSMENT
+      </Link>
+      <Link
+        href={
+          `/admin/course/${groupId}/assignment/${assignmentId}/statictics` as const
+        }
+        className={cn(
+          'flex h-full w-1/4 items-center justify-center rounded-full',
+          isCurrentTab('statictics') &&
+            'text-primary border-primary border-2 border-solid bg-white'
+        )}
+      >
+        STATISTICS
+      </Link>
     </div>
   )
 }
